+++ app/subscriber/src/components/footer/styled/SubscriberFooter.tsx
