+++ app/subscriber/src/components/footer/styled/index.ts
@@ -1,0 +1,1 @@
+export * from './SubscriberFooter';
