--- app/subscriber/src/features/home/index.ts
@@ -1,5 +1,0 @@
-export * from './InfoPanel';
-export * from './InfoText';
-export * from './Landing';
-export * from './LogoPanel';
-export * from './UnauthenticatedHome';
