--- app/subscriber/src/features/home/login/styled/index.ts
@@ -1,2 +1,0 @@
-export * from './BrowserLogin';
-export * from './MobileLogin';
